word_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
nltk==3.9.1
wordfreq==3.1
python-dotenv==1.0.1